January 30 changes
Primarily, we added left and right paddles and added the ability to move the left paddle.  We did this by creating variables leftY and rightY that control the paddles' horizontal position.  We changed the behavior when the up and down keys are pressed, to increase and decrease leftY and rightY.